统一Rpc帮助类
将所有Rpc统一用RpcHelper处理，客户端，服务器统一继承Rpc接口

+++ Component/UnityComponent/NetworkSceneManager.cs
@@ -160,7 +160,7 @@ namespace Net.UnityComponent
                 return;
             if (!identity.isLocal)
                 return;
-            identity.syncVarInfos[opt.index1].value = default;
+            identity.syncVarInfos[(ushort)opt.index1].value = default;
         }
 
         /// <summary>
